Create custom elements when the runtime finishes loading
